requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx>=0.26,<0.28
beautifulsoup4==4.12.3
lxml==5.3.0
supabase==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
apscheduler==3.10.4
aiosmtplib==3.0.2
aiofiles==25.1.0
email-validator==2.2.0
python-dotenv==1.0.1
pydantic[email]==2.10.4
pydantic-settings==2.7.0
selectolax==0.3.27
tenacity==9.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
